Move Mageia to DistributionInstaller
